float64

+++ utils.py
@@ -158,12 +158,12 @@ def coin_loss(_alphas, _betas, epsilon):
             if (replace_alphas[i]):
                 alpha = tf.stop_gradient(coin_coef_a(y_true[:, :, :, i], y_pred[:, :, :, i], epsilon))
             else:
-                alpha = tf.stop_gradient(coin_coef_a(tf.random.uniform(y_true[:, :, :, i].shape, 0, 1), tf.random.uniform(y_true[:, :, :, i].shape, 0, 1), epsilon)) # float(alphas[i])
+                alpha = tf.stop_gradient(coin_coef_a(tf.random.uniform(y_true[:, :, :, i].shape, 0, 1, dtype=tf.float64), tf.random.uniform(y_true[:, :, :, i].shape, 0, 1, dtype=tf.float64), epsilon)) # float(alphas[i])
 
             if (replace_betas[i]):
                 beta = tf.stop_gradient(coin_coef_b(y_true[:, :, :, i], y_pred[:, :, :, i], epsilon))
             else:
-                beta = tf.stop_gradient(coin_coef_b(tf.random.uniform(y_true[:, :, :, i].shape, 0, 1), tf.random.uniform(y_true[:, :, :, i].shape, 0, 1), epsilon)) # float(betas[i])
+                beta = tf.stop_gradient(coin_coef_b(tf.random.uniform(y_true[:, :, :, i].shape, 0, 1, dtype=tf.float64), tf.random.uniform(y_true[:, :, :, i].shape, 0, 1, dtype=tf.float64), epsilon)) # float(betas[i])
 
             loss += (1 + K.sum(alpha * y_true[:, :, :, i] * y_pred[:, :, :, i] + beta * y_pred[:, :, :, i]))
             iter += 1
